+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
-- [Patch] Updated ** ManagerSideNav** to allow a node for projectName prop ([#1428](https://github.com/optimizely/oui/pull/1428))
+- [Patch] Updated **ManagerSideNav** to allow a node for projectName prop ([#1428](https://github.com/optimizely/oui/pull/1428))
 
 ## 48.4.2 - 2020-10-02
 - [Patch] Updated **Tile** with new `unsavedChangesText` prop to indicate a changed item ([#1427](https://github.com/optimizely/oui/pull/1427))